+++ cmd/my-cli/root.go
@@ -51,8 +51,8 @@ Run "my-cli <command> --help" for details on a specific command.`,
 	rootCmd.AddCommand(newOpenaiChatTestCmd())
 	rootCmd.AddCommand(newFalconCisExportCmd())
 	rootCmd.AddCommand(newNexusExportCmd())
-	rootCmd.AddCommand(newIPLookupCmd())
-
+        rootCmd.AddCommand(newIPLookupCmd())
+        rootCmd.AddCommand(newHTTPStaticServerCmd())
 	return rootCmd
 }